requirements.txt
mlflow~=2.20.2
neptune~=1.13.0
python-dotenv~=1.1.0
numpy~=2.2.1
matplotlib~=3.10.0
seaborn~=0.13.2
pandas~=2.2.3
easydict~=1.13
pillow~=11.1.0
tqdm~=4.67.1
opencv-python~=4.11.0.86
torchinfo~=1.8.0